AUTH MUTATION ADDED

+++ app/checkconfigfile.go
@@ -171,7 +171,7 @@ func IsConfigurationOk(v *viper.Viper, silent bool, logToGlog bool) bool {
 	}
 
 	if v.IsSet("auth.tokenDefaultTimeout") {
-		if v.GetUint32("auth.tokenDefaultTimeout") == 0 {
+		if v.GetInt64("auth.tokenDefaultTimeout") == 0 {
 			SpitError("auth.tokenDefaultTimeout can not be 0", silent, logToGlog)
 			return false
 		}
@@ -181,7 +181,7 @@ func IsConfigurationOk(v *viper.Viper, silent bool, logToGlog bool) bool {
 	}
 
 	if v.IsSet("auth.tokenSuperGroupTimeout") {
-		if v.GetUint32("auth.tokenSuperGroupTimeout") == 0 {
+		if v.GetInt64("auth.tokenSuperGroupTimeout") == 0 {
 			SpitError("auth.tokenSuperGroupTimeout can not be 0", silent, logToGlog)
 			return false
 		}
@@ -191,7 +191,7 @@ func IsConfigurationOk(v *viper.Viper, silent bool, logToGlog bool) bool {
 	}
 
 	if v.IsSet("auth.tokenNormalGroupTimeout") {
-		if v.GetUint32("auth.tokenNormalGroupTimeout") == 0 {
+		if v.GetInt64("auth.tokenNormalGroupTimeout") == 0 {
 			SpitError("auth.tokenNormalGroupTimeout can not be 0", silent, logToGlog)
 			return false
 		}
@@ -201,7 +201,7 @@ func IsConfigurationOk(v *viper.Viper, silent bool, logToGlog bool) bool {
 	}
 
 	if v.IsSet("auth.tokenServiceGroupTimeout") {
-		if v.GetUint32("auth.tokenServiceGroupTimeout") == 0 {
+		if v.GetInt64("auth.tokenServiceGroupTimeout") == 0 {
 			SpitError("auth.tokenServiceGroupTimeout can not be 0", silent, logToGlog)
 			return false
 		}
